Cleaned up ODataQuery implementation
