+++ src/sdk/PnP.Core/QueryModel/Enums/FilteringConcatOperator.cs
@@ -1,0 +1,17 @@
+﻿namespace PnP.Core.QueryModel.Enums
+{
+    /// <summary>
+    /// Enumeration of logical concat operators for queries
+    /// </summary>
+    public enum FilteringConcatOperator
+    {
+        /// <summary>
+        /// Logical AND for query items in query groups
+        /// </summary>
+        AND,
+        /// <summary>
+        /// Logical OR for query items in query groups
+        /// </summary>
+        OR
+    }
+}
