+++ src/sdk/PnP.Core/QueryModel/Enums/OrderByDirection.cs
@@ -1,0 +1,17 @@
+﻿namespace PnP.Core.QueryModel.Enums
+{
+    /// <summary>
+    /// Enumeration of the ordering criteria for sorting results
+    /// </summary>
+    public enum OrderByDirection
+    {
+        /// <summary>
+        /// Ascending sorting
+        /// </summary>
+        Asc,
+        /// <summary>
+        /// Descending sorting
+        /// </summary>
+        Desc
+    }
+}
