+++ src/sdk/PnP.Core/QueryModel/Query/DataModelQueryService.cs
@@ -11,6 +11,7 @@ using System.Linq;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
+using PnP.Core.QueryModel.Enums;
 
 namespace PnP.Core.QueryModel
 {
